Added the new Textbook Flashcard System.
There is now a game with different decks of flashcards for each of the
elememtary school textbooks.
Let's Try 1
Let's Try 2
New Horizon 1
New Horizon 2

The decks are organized by unit and can be reviewed in order or at random.
The user can also select which cards from the deck to include in the review.

There is an auto mode which progresses through the cards on a timer, and
the user can hide the japanese or image attached to the cards as well.

Currently only the decks for Let's Try 1 are included.

Future additions will include games, such as "What's missing", "NG Word Game", and maybe even the Memory Game.