Build: (e8f5f07) Added a new Matching/Memory game! Due to user request.
Can toggle image or English on and off.
Can set how many sets of cards are displayed (3 to 13 sets / 6 to 26 cards).
Has an Auto Flip option where cards flip back over if not matching. Can be turned off.
Added letters to the back side of the cards for easier call outs.

Updated the settings menu for the bomb game to be less ugly.

Added a simple Merry Christmas message to the dashboard.